Feature/num 2303 delete attachments when delete project

+++ src/main/java/de/vitagroup/num/service/ProjectService.java
@@ -142,6 +142,7 @@ public class ProjectService {
 
         if (project.isDeletable()) {
             projectRepository.deleteById(projectId);
+            attachmentService.deleteAllProjectAttachments(projectId, userId);
             log.info("Project {} was deleted by {}", projectId, userId);
         } else {
             throw new ForbiddenException(ProjectService.class, CANNOT_DELETE_PROJECT_INVALID_STATUS,
@@ -679,7 +680,7 @@ public class ProjectService {
 
     private void deleteAttachments(Set<Long> attachmentsToBeDeleted, List<String> roles, UserDetails user, Project projectToEdit) {
         if (CollectionUtils.isNotEmpty(attachmentsToBeDeleted)) {
-            if (Roles.isProjectLead(roles)) {
+            if (Roles.isProjectLead(roles) && projectToEdit.isCoordinator(user.getUserId())) {
                 if (ProjectStatus.DRAFT.equals(projectToEdit.getStatus()) || ProjectStatus.CHANGE_REQUEST.equals(projectToEdit.getStatus())) {
                     attachmentService.deleteAttachments(attachmentsToBeDeleted, projectToEdit.getId(), user.getUserId(), false);
                     log.info("Project lead {} removed attachments {} from project {}", user.getUserId(), attachmentsToBeDeleted, projectToEdit.getId());
